+++ src/geometry/direction.rs
@@ -1,6 +1,7 @@
 /// Direction enum
-#[derive(Debug, PartialEq, Clone, Copy)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, Default)]
 pub enum Direction {
+    #[default]
     Vertical,
     Horizontal,
 }